UI窗口实现

+++ FourOperations/FourOperations/BaseData/BasicData.cs
@@ -10,7 +10,7 @@ namespace ConsoleApp2
         public int LsitNumA { get; set; }
         public int LsitNumB { get; set; }
         public string Theme { get; set; }//题目
-        public decimal Answer { get; set; }//答案
+        public double Answer { get; set; }//答案
         protected string FFF { get; set; } = "=";
 
         public List<string> CCC = new List<string> { "+", "-", "*", "/" };
